Move methods to PipelineBase
